+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## (5.2.0 IN PROGRESS)
 
 * Align the `finance.fund` interface version (`3.0`). Refs UIPFPOL-68.
+* Update the order lines query with location filtering to include the `searchLocationIds` field. Refs UIPFPOL-72.
 * Add additional order filters to support reporting requirements. Refs UIPFPOL-71.
 
 ## [5.1.0](https://github.com/folio-org/ui-plugin-find-po-line/tree/v5.1.0) (2024-03-19)